asfsf

+++ ft_strlen.c
@@ -1,7 +1,8 @@
 #include "libft.h"
-int	ft_strlen(const char *str)
+
+size_t	ft_strlen(const char *str)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (*str)
